Some changes

+++ QuickCourses.Api.Data/Repositories/RepositoryBase.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using MongoDB.Bson;
 using MongoDB.Driver;
@@ -15,6 +16,11 @@ namespace QuickCourses.Api.Data.Repositories
 
         public RepositoryBase(Settings settings)
         {
+            if (settings == null)
+            {
+                throw new ArgumentNullException(nameof(settings));
+            }
+
             Context = new Context<TValue>(settings);
         }
         
@@ -26,23 +32,48 @@ namespace QuickCourses.Api.Data.Repositories
 
         public virtual async Task<TValue> GetAsync(string id)
         {
+            if (id == null)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
             var result = await Context.Collection.Find(value => value.Id == id).FirstOrDefaultAsync();
             return result;
         }
 
         public virtual async Task<bool> ContainsAsync(string id)
         {
+            if (id == null)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
             var value = await GetAsync(id);
             return value != null;
         }
 
         public virtual async Task ReplaceAsync(string id, TValue newValue)
         {
+            if (id == null)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
+            if (newValue == null)
+            {
+                throw new ArgumentNullException(nameof(newValue));
+            }
+
             await Context.Collection.ReplaceOneAsync(value => value.Id == id, newValue);
         }
 
         public virtual async Task<string> InsertAsync(TValue value)
         {
+            if (value == null)
+            {
+                throw new ArgumentNullException(nameof(value));
+            }
+
             var result = ObjectId.GenerateNewId().ToString();
             await Context.Collection.InsertOneAsync(value);
             return result;
@@ -50,6 +81,11 @@ namespace QuickCourses.Api.Data.Repositories
 
         public virtual async Task<bool> DeleteAsync(string id)
         {
+            if (id == null)
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
             var result = await Context.Collection.DeleteOneAsync(value => value.Id == id);
             return result.DeletedCount == 1;
         }
